Refactors module loading and parsing, adds arch bindings
Adds a macro for exporting C API functions and improves
configuration and documentation.

+++ kmod-loader/src/arch/x86_64.rs
@@ -116,7 +116,7 @@ impl X86_64RelocationType {
             // Write the relocated value
             match size {
                 4 => location.write::<u32>(target_addr as u32),
-                8 => location.write::<u64>(target_addr as u64),
+                8 => location.write::<u64>(target_addr),
                 _ => unreachable!(),
             }
         }
